extract OCI package
