+++ pkg/fleet/installer/oci/mirror.go
